projectFileSystem/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
charset-normalizer>=3.3.2
langchain-openai>=0.1.0
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1